feat: specialty config system, per-specialty triage/assessment factories
- Specialty config Pinia store with ETag caching
- VitalFieldRenderer component (bp/numeric/select dispatch)
- TriageFactory + GeneralTriageSection + PediatricsTriageSection + FMTriageSection
- AssessmentFactory + GeneralAssessmentSection + per-specialty placeholders
- PatientSpecialtyFactory + placeholder sections (FM/Peds)
- Unified vitals pipeline (splitVitalsForApi helper)
- Auto-set specialty on consultation create
- Specialty picker in CredentialsForm (replaces freeform text)
- TriageTab and AssessmentTab now thin wrappers over factories
