+++ src/domains/patient/components/specialties/general/GeneralPatientSections.vue
@@ -1,0 +1,7 @@
+<script setup lang="ts">
+// General specialty — no extra patient sections needed
+</script>
+
+<template>
+  <!-- No specialty-specific sections for general practice -->
+</template>
